Refactor: Rename components to align with CRUD functionality
