+++ src/components/products/ProductSearchBar.vue
